and one more time

+++ tests/20_ll_update.sql
@@ -180,7 +180,7 @@ $v$ values(1) $v$
 );
 
 
-select results_eq($q$select count(*) from bi_temp_tables.devices 
+select results_eq($q$select count(*)::integer from bi_temp_tables.devices 
 where device_id=11 
 and  '[2016-03-16,  2016-03-16]'<@ effective 
 and '[2016-02-03, 2016-02-03]' <@ asserted $q$, 
@@ -196,7 +196,7 @@ select results_eq($q$select * from bitemporal_internal.ll_bitemporal_delete('bi_
 $v$ values(2) $v$
 ,'bitemporal delete');
 
-select results_eq($q$select count(*) from bi_temp_tables.devices 
+select results_eq($q$select count(*)::integer from bi_temp_tables.devices 
 where device_id=1 
 and '[2016-04-05, 2016-04-05]' <@ asserted $q$, 
 $v$ values(0::integer) $v$
